Make reinforcement drawing view object parametric
- Change object name from ReinforcementDrawingContent to ReinforcementDrawingView
- Add structure and Rebars parameter to ReinforcementDrawingView object,
  which keeps generated svg content in sync with rebars in model
- Some code improvements
